refactor: integrate icons with routes and extract NavigationMenuItem
- Move navigationIcons from NavigationMenu to routes.tsx RouteInfo
- Add icon property to RouteInfo interface with ComponentType<SvgIconProps>
- Extract NavigationMenuItem component from NavigationMenu for better modularity
- Replace primaryTypographyProps with Typography children for cleaner JSX
- Reduce NavigationMenu from 78 to 45 lines (43% reduction)
- Improve component separation following single responsibility principle
